+++ modules/components/item/Group.jsx
@@ -1,12 +1,12 @@
 import React, { Component, PureComponent } from "react";
 import PropTypes from "prop-types";
 import startsWith from "lodash/startsWith";
-import GroupContainer from "./containers/GroupContainer";
-import Draggable from "./containers/Draggable";
+import GroupContainer from "../containers/GroupContainer";
+import Draggable from "../containers/Draggable";
 const classNames = require("classnames");
 import Item from "./Item";
 import {GroupActions} from "./GroupActions";
-import {ConfirmFn, DragIcon, dummyFn} from "./utils";
+import {ConfirmFn, DragIcon, dummyFn} from "../utils";
 
 const defaultPosition = "topRight";
