+++ modules/components/item/RuleGroup.jsx
@@ -1,12 +1,12 @@
 import React from "react";
 import PropTypes from "prop-types";
-import GroupContainer from "./containers/GroupContainer";
-import Draggable from "./containers/Draggable";
+import GroupContainer from "../containers/GroupContainer";
+import Draggable from "../containers/Draggable";
 import {BasicGroup} from "./Group";
 import {RuleGroupActions} from "./RuleGroupActions";
-import FieldWrapper from "./FieldWrapper";
-import {useOnPropsChanged} from "../utils/stuff";
-import {ConfirmFn} from "./utils";
+import FieldWrapper from "../rule/FieldWrapper";
+import {useOnPropsChanged} from "../../utils/stuff";
+import {ConfirmFn} from "../utils";
 
 
 @GroupContainer
