+++ modules/components/item/RuleGroupExt.jsx
@@ -1,13 +1,13 @@
 import React from "react";
 import PropTypes from "prop-types";
-import GroupContainer from "./containers/GroupContainer";
-import Draggable from "./containers/Draggable";
+import GroupContainer from "../containers/GroupContainer";
+import Draggable from "../containers/Draggable";
 import {BasicGroup} from "./Group";
 import {RuleGroupActions} from "./RuleGroupActions";
-import FieldWrapper from "./FieldWrapper";
-import OperatorWrapper from "./OperatorWrapper";
-import {useOnPropsChanged} from "../utils/stuff";
-import {ConfirmFn} from "./utils";
+import FieldWrapper from "../rule/FieldWrapper";
+import OperatorWrapper from "../rule/OperatorWrapper";
+import {useOnPropsChanged} from "../../utils/stuff";
+import {ConfirmFn} from "../utils";
 
 
 @GroupContainer
