+++ modules/components/widgets/index.js
@@ -1,6 +1,6 @@
 
-export {default as ValueFieldWidget} from "../ValueField";
-export {default as FuncWidget} from "../FuncWidget";
+export {default as ValueFieldWidget} from "../rule/ValueField";
+export {default as FuncWidget} from "../rule/FuncWidget";
 
 //export * from './antd/index.js';
 export * from "./vanilla/index.js";
